gif_generator/requirements.txt
Flask
imageio
numpy
Pillow
